requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Bootstrap==3.3.7.1
python-dotenv==1.0.1
requests==2.32.3
wtforms==3.1.2
email-validator==2.1.1
passlib[bcrypt]==1.7.4
